id:3470 file renames for forgotPassword* -> passwordManagement*
